+++ gui/browserWidgets/colormenu.cpp
@@ -26,7 +26,8 @@ ColorMenu::ColorMenu(QObject *parent) :
     QObject(parent)
 {
     this->parent = parent;
-    currentColor.setNamedColor("black");
+    setCurrentColor(Qt::black);
+
     populateList();
     QString css = global.getThemeCss("colorMenuCss");
     if (css!="")
@@ -79,12 +80,12 @@ void ColorMenu::populateList() {
         menu.addAction(newAction);
         connect(newAction, SIGNAL(hovered()), this, SLOT(itemHovered()));
     }
-    QLOG_DEBUG() << "Done: populating colormenu";
+    //QLOG_DEBUG() << "Done: populating colormenu";
 }
 
 
 QColor *ColorMenu::getCurrentColor() {
-    QLOG_DEBUG() << "ColorMenu::getCurrentColor; currentColor=" << currentColor.name();
+    //QLOG_DEBUG() << "ColorMenu::getCurrentColor; currentColor=" << currentColor.name();
     return &currentColor;
 }
 
@@ -95,7 +96,7 @@ QString ColorMenu::getCurrentColorName() {
     if (colorName.isEmpty()) {
         colorName = colorCode;
     }
-    QLOG_DEBUG() << "ColorMenu::getCurrentColorName; currentColorName=" << colorName;
+    //QLOG_DEBUG() << "ColorMenu::getCurrentColorName; currentColorName=" << colorName;
     return colorName;
 }
 
@@ -103,9 +104,7 @@ QMenu *ColorMenu::getMenu() {
     return &menu;
 }
 
-
-
-void ColorMenu::setDefault(QColor color) {
+void ColorMenu::setCurrentColor(QColor color) {
     currentColor = color;
 }