+++ gui/browserWidgets/ntitleeditor.h
@@ -55,6 +55,7 @@ signals:
   void focussed(bool hasFocus);
   void titleChanged();
   void titleUpdated(QString);
+  void setMessage(QString msg);
 
 protected:
   virtual void focusInEvent(QFocusEvent *e);