+++ docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 2018-06
 
+### #5
+* Minor improvement in font color/background color (highlight selection). 
+  Tooltip now show how to use. Keyboard shortcuts possible (default Ctrl-d, Ctrl-Shift-H). Context
+  menu option added.
+
 ### #4
 * Merged changes from [jeffkowalski/Nixnote2](https://github.com/jeffkowalski/Nixnote2) - **various fixes, build config for macOS**
 * **Rewrite of the path handling logic after merge. Now standard compliant path for config and user data are used.**